Split required methods from ms_icpr library
