+++ lib/msf/core/exploit/remote/http/web_enrollment.rb
@@ -4,10 +4,13 @@ module Msf
       module HTTP
         # This module provides a way of interacting with the Microsoft AD/CS web enrollment portal
         module WebEnrollment
-          include Msf::Exploit::Remote::MsIcpr
+
+          include Msf::Exploit::Remote::MsCsr
+
           def initialize(info = {})
             super
-            deregister_options('CA', 'CERT_TEMPLATE', 'ADD_CERT_APP_POLICY', 'RPORT', 'SMBDomain', 'SMBPassword', 'HttpUsername', 'HttpPassword')
+            #deregister_options('CA', 'CERT_TEMPLATE', 'ADD_CERT_APP_POLICY', 'RPORT', 'SMBDomain', 'SMBPassword')
+            deregister_options('HttpUsername', 'HttpPassword')
             register_options([
                                Opt::RPORT(80),
                                OptString.new('HttpUsername', [false, 'The HTTP username to specify for authentication', '']),
@@ -28,7 +31,6 @@ module Msf
               algorithm: datastore['DigestAlgorithm'],
               application_policies: datastore['ADD_CERT_APP_POLICY']
             )
-            #vprint_status("CSR:\n#{csr.to_s}")
             vprint_status('CSR Generated')
             on_behalf_of = datastore['ON_BEHALF_OF'].blank? ? nil : datastore['ON_BEHALF_OF']
             if @pkcs12 && on_behalf_of
